% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_parts.R
\name{createFilteredDt}
\alias{createFilteredDt}
\title{create filtered dt}
\usage{
createFilteredDt(min_year, max_year)
}
\arguments{
\item{min_year}{min_year}

\item{max_year}{max_year}
}
\value{
filtered dt
}
\description{
create filtered dt
}
\examples{
\donttest{createFilteredDt(2007, 2015)}
}
